move pdl interpretation code to own file in xdsl module
